mod_operator: Add operator services module.
This adds an operator services module for Asterisk
queues and queue agents.
This module is intended for a specific use case,
but has been designed in a generic manner that enables
it to be adapted for a myriad of purposes, so it is
being included in the main source tree for reference
and maintainability. It also serves as an example of
implementation of queue handlers.

The sample config is omitted, and thus this module
will generally decline to load on most systems.