modman: Add module dependency checker utility.
Add an external program "modman", which is used by the build
system to check for missing dependencies and automatically
disable modules that cannot be built due to missing dependencies.
This will streamline building of modules when dependencies for
some of them are missing, and we just want to build the ones
that can be built with the libraries that are present.