node: Fix node interruption and ANSI toggling limitations.
If a node is buffered, the current node interrupt
technique does not interrupt the node until after
the buffered read returns. To make it immediate,
spoof input on the PTY master to cause the interrupt
to be processed immediately.

This also fixes PTY handling of ANSI support. Previously,
this was only checked when the PTY thread started; since
this can change at any point, check it at the beginning
of each loop iteration instead. When toggling ANSI on a node,
also reset colors and pause briefly to ensure these are
processed before any further output.