door_irc: Add color formatting and increase usability.
Add color formatting to the door_irc IRC client, to make the
client easier and more interesting to use. When establishing
a 1:1 connection (irc_single_client), present information in
a collated manner, rather than just dumping raw numerics.
Additionally, allow users to query a list of users in the channel
on-demand.