node.c: Add terminal autodetection capabilities.
For interactive sessions, try to determine if the client's
terminal supports ANSI or not and the speed at which the
client's link is operating. This allows us to be sure about
ANSI support, if we've autodetected it, and also allows us
to infer the client's speed without asking.

Also clean up the node CLI commands to only print certain
information for interactive nodes, to reduce confusion.