net_irc: Add private namespace channels to prevent relay collisions.
Add the ability for channels to exist in a "private"
per-user namespace, which allows a channel to be restricted
to a single user. This is mostly useful in conjunction
with private relays, to ensure that the channel is automatically
access restricted and also to prevent channel naming conflicts
between users/relays, since each such channel is unique by user,
thus avoiding naming collisions.

This extends the modeless channel prefix '+' for this purpose;
we still honor the modeless property of the channel, since modes
are not useful for private namespace channels.