node.c: Default terminal size to 80x24.
The default terminal size is 80 x 24, but this
is left zero-initialized, which means that dial-in
clients were only shown a single column in menus,
even though their terminal is probably 80 x 24.
Previously, this was left as 0 x 0 to distinguish
an explicit size from the default; for ease of use,
this is changed to 80 x 24 by default, and we use
a separate flag (node->dimensions) to keep track of
having received explicit dimensions.