lock.c: Add locking wrapper.
This adds a locking wrapper to the BBS core, to make it
easier to debug locking-related issues. This allows locking
correctness to be enforced more stringently (helping to
squash more bugs long-term) and triggers warnings if
deadlocks are detected. The performance overhead should be
pretty minimal, so this is more efficient embedded in this
manner as opposed to using a tool like helgrind.

This touches a sprawling amount of lock-using code, since
all mutexes and rwlocks have been updated to use the wrapper,
but in the future, any changes should be contained to lock.c
and lock.h, since API usage is the same.

Other related work included in this patch:

* We were never properly initialize non-static list heads
  in the linked lists API, so macros for init and destroy
  have been added, and appropriate calls sprinkled throughout
  the codebase.
* Fix deadlock in mod_slack caused by reusing the same lock
  for too many things (a callback could be triggered with
  the lock held, causing a deadlock).