net_irc: Allow private namespace prefix to be overridden.
It turns out that some (many? most?) IRC clients do not
properly support channels that do not begin with #, e.g.
Ambassador. As such, the default implementation of private
namespace channels is unusable with these clients. To be
more flexible, we now allow this to be easily changed in
net_irc.h.