door_irc: Fix sender prefix regression for relay.
A previous change refactored and cleaned up the code but also made it
so that a NULL sender was passed in for many cases, preventing us from
including the sender as part of the relayed message. Additionally, the
existing logic does not work properly, so the relayed messages now
included the sender as part of the message, rather than trying to use
it as the prefix, which won't work (have any effect) towards remote
IRC networks.