mod_operator: Improve reverse lookup option.
Prepopulate the calling number as the default,
so the agent doesn't need to manually key it in.
This can still be edited if needed.

Also add a default handler for queues that currently
don't have a handler, which contains the reverse lookup
option, since this could be useful for those queues.