socket.c: Return -1 in poll_read if read returns 0.
Currently, bbs_poll_read and bbs_node_poll_read both
return the return value from poll and read unmodified.
However, this can lead to callers misinterpreting the
return value when read returns 0, and this is treated
as a poll timeout, since these cannot be differentiated.

We could expand the return values to differentiate better
(e.g. return -2 for poll failure), but most existing usage
of these APIs already assume a return value of 0 is from
poll and treats negative return values as failure, so just
return -1 if read returns 0 to make everything "just work".
In particular, this now ensures that the SMTP server does not
try to send a "451 timeout" message when a client disconnects.