scripts: Add convenience script to disable wall messages from syslog.
If the server is being used as a syslog server, then it is prone to
receiving emergency messages from the outside that can trigger wall
messages to go to all the system terminals. This won't impact BBS
nodes at all, but can be bothersome to the sysop. Add a script to
disable this, for convenience.