logger.c: Fix log message truncation check.
Fix small typo that previously compared against
the size of a char*, instead of the size of the
log buffer. Because the latter was much larger,
this was safe (though incorrect); however, this
resulted in almost always assuming the message
was truncated in the stack-allocated buffer when
it was not and thus unnecessarily redoing the
formatting using a dynamic buffer instead. Thus,
the main implication of this bug fix is that performance
should improve a decent bit, since we now avoid
unnecessary allocations and formatting operations
for the vast majority of log messages.