test_imap_pop3_compat: Improve synchronization in flaky test.
Occasionally, test_imap_pop3_compat fails to run successfully
due to a race condition where the POP3 login will process
before the IMAP TESTLOCK command begins, resulting in the
opposite of the expected scenario (POP3 gaining an exclusive
lock and preventing IMAP access, rather than IMAP gaining a
a non-exclusive lock and preventing POP3's exclusive access).

This adds an untagged reply to the TESTLOCK command in net_imap
that the test suite can use to synchronize, to ensure the POP3
PASS command is not sent prior to the IMAP client acquiring
the mailbox lock.