README: Trim redundant info to work around GitHub rendering bug.
This removes an explicit enumeration of most BBS config files from
the config overview section. The main reason this is being done is
to work around a known GitHub bug that causes issues rendering large
RST files. Since the README is currently just over the threshold that
causes rendering to fail, this trims some fat so that the README will
render on GitHub's website again. However, the removed lines don't
contain any new information, so this is just as well.