nets, transfer.c: Use usernames for home directory names.
Currently, the file paths that are presented to users
are very literal - just a subdirectory of the actual paths
that are actually on disk. This is generally fine; however,
it's not ideal for home directories, where it makes sense
to usernames instead of user IDs, which is how the directories
are named on disk.

This adds some conversion logic inside transfer.c, and minimally
to net_ftp and net_ssh, where relevant, to do the needed conversions
between user IDs and usernames to make this happen.