net_ftp: Advertise correct IP for PASV depending on source.
Originally, PASV would autodetermine the IP to use by using the first
interface's address, which doesn't work if there are multiple
addresses, or if the client is behind NAT.

To fix this, we now iterate all the interfaces and pick the best one
based on if the netmask matches the client's address. Additionally,
an IP address for PASV can be explicitly specified, one for both
private and public IP ranges.