nets: Detect terminal type and speed.
For interactive terminal network protocols (Telnet, SSH, RLogin),
detect the terminal type and speed, if supported by the protocol,
so that TERM can be set appropriately when needed, like executing
an external program.

Also detect if a term type is in use that isn't in the terminfo
database, and add syncterm terminfo support to the server setup
script.

This also revamps the options negotiation in net_telnet to be more
robust and more compatible with different terminal emulators.