system.c: Support retaining network connectivity in container.
The existing isoexec option uses CLONE_NEWNET, so there is no
network connectivity inside the container. While ideal for some
applications, this currently forces use of the plain "exec"
handler for anything requiring the network, which is not ideal.

This adds the ability to retain network connectivity by omitting
the CLONE_NEWNET flag, and adds a new handler, "isonetexec",
that can be used for this purpose. The isoroot program also
supports the -n option to keep network, useful for installing
packages or maintenance activities.

While the new isonetexec handler can still be dangerous, since
it can allow unrestricted network connections (it is not filtered
in any way), it is still far safer than using the plain "exec"
handler, so this should give users some much-needed flexibility.

Additionally:

* We now also change directories to a user's directory when
  launching a container, which is a more reasonable place
  for programs to get launched.
* The "Press a key" prompt after running external programs
  using the exec, isoexec (and now isonetexec) handlers is
  suppressed if the program exited with a 0 return code.
  The prompt still appears if the program exited nonzero.