net_imap: Avoid desynchronization from unsolicited UID during MOVE/COPY.
When doing a transparent cross-server MOVE or COPY operation, we would
originally request only three items, not including the UID, since it
doesn't matter. However, some servers will send the UID anyways, and
in particular, if this was sent after all the requested items, our
parsing logic would have moved onto the next message and this would
result in reading unexpected data that didn't correspond to the next
message. This could be handled by having logic to detect receiving
the UID and ignore it, but a cleaner way is to simply always request
the UID and discard it, to avoid getting out of sync like this to
begin with.