mod_mail: Log remote IMAP server error when LOGIN fails.
Fix an off-nominal path to display the error sent by
a remote IMAP server (for IMAP client proxy) when
the IMAP LOGIN command fails. Previously, we were using
IMAP_CLIENT_EXPECT, which would be expected to fail and
thus goto the end of the function, rather than logging
the error first, making this a confusing error to follow.

This bug was initially discovered due to logins to Microsoft
accounts failing intermittently. This was initially presumed
to be a bug in capability parsing, but is actually (even by
their own admission) a bug in Microsoft's email servers:

"Users [using] Basic Authentication may experience recurring
password prompts... this is a known issue."

No workaround will be implemented to accomodate this since
this is a substantial (and, frankly, inexcusable for anything
calling itself a mail server) bug in Microsoft's software,
which should NOT be handled by other applications.

Additionally, Microsoft is disabling basic authentication in
three days (September 16, 2024), so this will be moot
very shortly. In any case, the LOGIN error sent by the server
is now logged properly to alert users to this happening.

LBBS-62 #close