scripts: Add basic setup wizard script.
This script was created more than a year ago to assist
new users with some basic configuration initialization
tasks. It is not very comprehensive, but configures
many of the basic settings required to get started
with running a basic BBS setup.

Much of the functionality of this script is redundant
considering what server_setup.sh already does, with
the main exception being the database configuration.
As such, this script is being committed "as is".