net_smtp: Don't send non-final 220 greeting to relaying clients.
Skip the non-final 220 response when handling connections
from hosts authorized to relay mail for some domain(s).
While this should not be necessary when dealing with
compliant SMTP clients, there is a lot of poorly written
SMTP client code in enterprise software that will choke
if receiving a non-final 220 response followed by a final
220 response, such as the HP Management Agent Event Notifier.
This "feature" is intended more for dealing with inbound
mail to an MTA from the Internet, so doesn't do much in
this scenario anyways; thus, for compatibility, we don't
do this anymore when dealing with relays.