mod_webmail: Properly handle high-latency responses.
If messages have been expunged from a remote mailbox since net_imap
last cached their size, it will have to issue a FETCH 1:* to get the
SIZE of every message. This can take a long time on large mailboxes,
and occasionally this can take long enough that libetpan times out
on the LIST command and returns.

Previously, this would mess up our synchronization such that when the
end of the LIST response was finally received, it was treated as
normal unsolicited untagged responses, such as during IDLE. This was
because we weren't actually checking if LIST succeeded and thus we
continued on normally. This triggered an assertion when the untagged
responses for the end of the LIST response arrived, as we were also
erroneously attempting to restart IDLE even if we weren't idling prior
to the poll data callback firing.

These issues have been addressed as follows:

* Don't attempt to restart IDLE if we weren't idling prior to the
  on_poll_activity callback firing, since this is wrong and causes
  an assertion.
* Return success or failure from list_response so we can abort
  altogether if the LIST command does happen to fail for some reason.
* Use a larger timeout for reading data in the LIST response, to
  account for it potentially taking a while. Instead of 15 seconds,
  a 45 second timeout is now used for STATUS and LIST, since these
  commands can trigger a FETCH 1:* behind the scenes. This way,
  the LIST command doesn't fail simply because it took longer than
  the default timeout to get the size of a mailbox.

LBBS-12 #close