modman: Create dummy .so targets last, not first. (#42)
On non-Debian systems, modconfig is frequently unsuccessful
in creating dummy build targets that will cause make to
ignore a target and not build it at all. This is due to
modman creating the dummy .so target first, followed by
the .d target. This can occasionally lead to make attempting to build
targets we wanted to be ignored, causing a build failure.

To avoid this, create the .so target last.