crypt.c: Don't attempt to seed crypt_data.current_salt.
current_salt is a deprecated field that is only set when
NEED_CRYPTO_IMPL is defined, such as on Debian 10 and
other older distros that don't provide a sufficiently
capable crypt_r. However, some distros, such as Rocky Linux 8,
don't even have these fields in their struct crypt_data.
Since it's not clear what this accomplished and appears to
be unnecessary, just disable this for now.

To ensure this doesn't cause a regression, all the Linux
CI's also run the menu tests, to ensure that authentication
succeeds (since that can rely on the affected code).