pty.c: Eliminate improper handling in PTY master relay thread.
The implementation of the PTY master thread has always been somewhat
kludged, but there are a few major issues with it that have caused some
serious problems to date.

First off, logic was initially added to look for ASCII ETX (3) manually
and send a signal to the child process of a node, if there was one.
This logic is totally and completely unnecessary. The kernel already
handles sending signals to the running process for us, and it's best
to just let the kernel do it. For example, if ssh is launched, we want
ssh to handle the SIGINT to send a SIGINT to the process running
remotely, rather than just blindly sending a signal to ssh ourselves.
This logic has been removed, with the only caveat now being that if the
running program is not well-behaved, it may not be as straightforward
to exit it, but this is how a real terminal emulator would handle it.
So, simply relay these bytes as is to the PTY master, without doing
anything special when we see them.

Secondly, there has been a lot of embedded logic added to pty.c over
time to handle the various sequences when a user presses "ENTER" on
a variety of terminal emulators and terminal devices (e.g. TDDs). Some
of this logic is unnecessary in the first place, such as the logic to
intercept a lone CR and convert it to CR LF - setting ICRNL will handle
this for us, provided that IGNCR is NOT set (and in the past, it was,
which is likely why this didn't work properly on its own, and a
workaround was added to pty.c). This has been removed.

Other protocol-specific logic has been embedded here over time, such
as the handling of CR NUL to treat it as CR LF, for Telnet. Now that
ICRNL works properly, we can simply treat CR NUL as CR, as feeding
that into the PTY master will yield a LF on the other side (in
normal canonical mode). This logic should ideally be in net_telnet,
but for now, it remains in pty.c, but only for Telnet connections.

The remaining line ending logic in pty.c has largely been removed as
it was incorrect and caused problems. For example, hitting ENTER
might be expected to just send ^M, but due to logic to turn CR into
CR LF, would actually send ^M ^J, and certain applications like
nano would not handle this well at all. Removing this unnecessary
and incorrect logic makes things behave as expected, since applications
will be able to handle these sequences as they desire, rather than
us twiddling line endings in the PTY layer where we frankly have no
business doing such things. The remaining logic that is actually
needed has been cleaned up and clearly separated out as protocol-specific.

As part of this change, some disabled-by-default code has been added
to pty.c to dump the input to a file, to be able to easily capture
sequences sent by various clients for use in automated tests simulating
these clients, which can help in constructing automated tests for
various terminal emulators.