net_imap: Support obsolete FETCH requests for part number 0.
IMAP4rev1 obsoleted the ability to request part number 0 in FETCH
for getting the RFC822 header, but some clients will still use
this syntax, typically as a fallback. Support this, but add a
warning, since if a client is using this format, it is likely due
to a bug somewhere.