net_ssh: Handle SSH session termination more cleanly and sanely.
Previously, ssh_channel_close would sometimes hang when
called at session termination. To work around this, preliminary
changes were made to set the channel to nonblocking mode and
try to cleanly end the session using SSH EOF. This can lead
to busy loops if done simplistically, since the client may
not end the session cleanly. As a compromise, try to do so
if we can, but use a loose loop and give up after a minute or so.

LBBS-57 #close