Makefile: Allow clobbering of non-PHP scripts.
At some point, the -n (no clobber) flag was added when copying
scripts into the installed scripts directory, but this was really
only intended to prevent clobbering of customizations made to the
sample IRC bot script, not the majority of scripts in general.
The no clobber flag prevented updated versions of shell scripts with
bug fixes from getting installed, so instead allow clobbering of
all script types in use except for PHP scripts.

It remains that many of the shell scripts in the scripts directory
are not particularly useful directly outside of the compilation
process, but for now, no change is being made to prevent any of these
scripts from being installed in the first place.