net_imap: Fix misordering when traversing a maildir's new dir.
Previously, traversals of a maildir's cur dir and new dir both used
maildir_ordered_traverse, which uses uidsort to sort the directory
entries. However, this is invalid when traversing a maildir's new dir,
as no UIDs have been assigned; this leads to an arbitrary an invalid
sorting since uidsort assumes both inputs have a UID in the filename
in order to sort them. This led to the order in which messages were
added to the new dir not being preserved when assigned UIDs.

Instead, we should just be using the entire filename by itself for
sorting, since the filenames (when ordered) are already in the right
order to assigning UIDs. So, use alphasort when traversing a new dir,
which preserves the order in which messages were added to the mailbox.