bbs: Ignore SIGHUP signal and don't skip logging if daemonized. (#47)
When running as a daemon, the BBS receives a SIGHUP whenever a remote
sysop console disconnects. Without handling the signal, this causes
the BBS to get killed by the kernel. To avoid this, install a dummy
signal handler when daemonized to ignore the signal and continue
running.

Additionally, logger.c was using the wrong check to determine whether
to generate a log message for outputting to terminals, which resulted
in no log messages on consoles when the BBS was daemonized. Adjust
this logic so that console messages are generated when needed.

Resolves: #46