mod_smtp_delivery_external: Retry plaintext delivery if STARTTLS fails.
Unless requirestarttls is true, we should reattempt delivery if
we attempted to use STARTTLS but it failed (e.g. certificate expired,
a self-signed certificate is being used, etc.) This is essentially the
same as if STARTTLS wasn't supported by the destination server in the
first place. This ensures that delivery doesn't fail purely due to the
receiving mail server having a certificate issue, unless, of course,
requirestarttls is enabled.

LBBS-75 #close