net_telnet: Implement RFC 1143 loop avoidance.
This implements option negotiation per the Q Method of Option
Negotiation, specified in RFC 1143. This helps ensure compliance
with the loop avoidance directives in RFC 854, and makes the
negotiation more event-driven. A few exceptions that violate
RFC 1143 are retained (and explicitly noted as such) to ensure
compatibility with certain clients. An unnecessary command has
also been removed since it did not comply with RFC 1143.