bbs.c: Fix restarts failing due to SIGHUP.
Use SIG_IGN as the signal handler for SIGHUP. Previously, a custom
handler (that did nothing) was used, which would cause restarts
to fail randomly, because because custom handlers are not preserved
by execve. SIG_IGN does not have this problem, since SIG_IGN is
preserved by execve, so by using it, we ensure that SIGHUP can't
be received during a restart and cause the BBS to exit.

Additionally, switch all signal handler setup to use sigaction(2),
rather than the obsolete signal(2), since there was no good reason
we were using it in the first place.