net_telnet: Fix flaky test_terminals test.
Because test_terminals does not emulate a real Telnet client and instead
expects responses to be in a consistent order, it occasionally fails
if responses come back in a different order than expected, due to being
split across multiple reads and not continuing to read if we haven't
processed all outstanding requests. Though this doesn't cause an issue
for real clients, to prevent test failures, we now ensure the NAWS
option has been fully processed before trying to negotiate the TSPEED
option. This ensures that TSPEED is not offered before NAWS has
finished, preserving the expected order for the test.

LBBS-77 #close