mod_smtp_filter: Fix privacy preservation for outgoing mail submissions.
For message submissions that are addressed to external recipients, the
processing for these messages was using the relay handler to prepend
the Received header. This handler would not respect submission privacy
configuration, which could result in the sender's IP address being leaked.
A failsafe check is added here to ensure that submissions have this
header added in a way that respects the current privacy setting.

As part of this change, logging using smtp_filter_write will not use
the calling location for this log message, rather than the location
inside this function, to make tracing the source of prepends easier.