net_imap: Reject commands if EXPUNGEs are pending.
Since net_imap deviates from IMAP4rev1 in that we maintain no
per-connection mapping of sequence numbers to UIDs from the
client perspective, it is not safe to process commands when
EXPUNGEs are pending of which to inform the client. Explicitly
reject such commands due to being out of synchronization, to
prevent an operation from being performed against a different
message than the client intended by the provided sequence number.
Although this is still not compliant, it prevents any operations
from being carried out that would violate the standard, and in
practice, it would be quite rare for this to actually happen.

If in the future, a mapping is added to maintain this
per-connection mapping, this could be refined.