io_tls: Fix blocked writes causing other TLS session I/O to get blocked.
Fix "deadlock" caused when doing a large COPY/MOVE operation
between remote IMAP servers, which involves doing a large FETCH
from one and then a series of APPENDs to the other. Since the TLS
thread only services one session at a time, the FETCH would saturate
the bbs_readline buffer and cause bbs_timed_write to wait 15 seconds
for the file descriptor to become writable (which it never would).
In the meantime, no other TLS sessions could be serviced, and this
would lead to breakage such as failing to receive a successful IMAP
reply from the destination. While the issue is more general than IMAP,
this exemplifies a dependency scenario where we are waiting for data
from session B before continuing to process session A, but the TLS
thread is retrying writes for session A and thus can't service session B.
To fix this, we now wait just 500 ms, rather than 15 seconds, before
saving the data for a future iteration and moving on. This ensures
that no individual session can be "starved" for too long and that
all sessions can continue to make progress.