io_log: Add I/O transformer that logs I/O sessions to a file.
Add I/O logging module that simply logs an I/O session to a file
for further analysis, similar in format to tcpflow. This is
useful for communication that is also using other transformations,
e.g. TLS, since tcpdump/tcpflow are not useful in these scenarios;
the application needs to be provide a mechanism to dump the I/O
prior to encryption (or vice versa, after decryption). This
transformation is never active by default.

As part of this, the I/O APIs have been extended in a few ways
to allow dynamically adding a transformation from the CLI.

Related fixes:
* module.c: Don't attempt to destroy stringlists if they were never
  initialized.