net_imap: Fix duplicated/malformed RFC822.HEADER FETCH response.
The handling for RFC822.HEADER previously resulted in output being
flushed immediately, before the response for that message even began.
Move it to afterwards, and also eliminate duplicate handling of this
FETCH item.

Additionally, send_message did not correctly handle the HEADERS_ONLY
case, as it would have the length of the headers but the offset of
the body start, rather than the headers start. This would result in
a request to bbs_sendfile to send past the end of the file if the
message's body was short enough, resulting in an infinite loop since
sendfile would return 0. To address this, bbs_sendfile now handles
a return of 0 by waiting for the file descriptor to become writable,
and also checks if the offset exceeds the file size. send_message
is also fixed to correctly handle HEADERS_ONLY.

A test has been added to verify both of the above.

LBBS-85 #close