module.c: Rewrite autoloading using a linked list.
This largely rewrites the module traversal logic of the module autoloading
that occurs at BBS startup. Previously, this used a set of string lists
to keep track of modules that should be preloaded, not loaded, required,
etc. Over time, this became very clunky and inefficient, and didn't allow
capturing relationships involving multi-layer dependencies at all.

Instead of this, we now use a single (doubly) linked list so that the
modules can be ordered correctly prior to actually autoloading them.
This also reduces the number of traversals order all and is more
efficient. Some logging has also been cleaned up to be less cluttered
and more usable for the user.

LBBS-2 #close