net_ws: Fix possible reference count leak.
It was possible for module references to be leaked by net_ws,
since find_route increases the module refcount (e.g. of mod_webmail),
but on certain off-nominal failure paths, the module refcount was
never decremented since we would jump to an exit path further in
the cleanup process. These have been adjusted so that all jumps
after find_route will clean up the previously leaked refcount.

This is likely the underlying issue that prompted LBBS-3.