config.c: Remove superfluous file open.
The current config file is opened before we begin the process
of reading it and copying over its contents in order to update
or insert a new configuration value, but this opened file is
never used, simply closed after this process completes. I
probably wrote it this way with the assumption the copy process
uses this file handle, but it doesn't; it opens its own. Therefore,
this file was opened and closed for no reason, and this code can
simply be removed. Doing so also seems to prevent this from being
leaked.