net_smtp: Allow mail submissions for non-local domains.
Previously, users were only able to submit mail for identites
that corresponded to local mailboxes on the system. However,
it may be desirable for a user to submit mail locally and have
it relayed to an authorized sender for this domain. This
adds the concept of "authorized identities" for local mail users,
which allows sending as any arbitrary identities, particularly
useful for multi-node private mail transit.