lock.c: Don't assert if MAX_READERS is exceeded.
This was initially done to draw attention to an unusually high number
of readers, but this can happen legitimately, such as if we get a
large number of connections at the same time, causing a bunch of
events to be generated (potentially more than MAX_READERS). While
this may or may not be malicious, there's nothing wrong on our side,
so asserting is inappropriate, just log a warning only.