mod_mailscript: Make EXEC action always safe to execute.
Originally, EXEC always executed programs on the host system,
which made the action, and the MailScript rule engine as a whole,
unsafe to allow users to directly control.

Now, EXEC only executes programs on the host for global
rules (which can only be set by the sysop). If found in
a mailbox rule, the program will be executed in an isolated
execution environment (same as isoexec, so in the container
and without networking). This allows users the flexibility
to use the ACTION, while limiting what they can do to things
they are already able to do anyways.

Because this now makes MailScript safe for users to edit directly,
the mailbox rules files are allowed to exist in both the maildir,
as before, as well as ~/.config/.rules (within the user's home directory).
The maildir location is retained since it is the only version for
non-user associated mailboxes (e.g. shared mailboxes). The user version
is useful since it is the only one the user can modify. Both versions
are executed if they exist.

Similarly, Sieve rules may exist either in the maildir or home
directory, but not both, unlike with MailScript. This is to avoid
complications within the ManageSieve protocol. For personal mailboxes,
Sieve scripts always exist in the home directory, which allows them to
now be edited within the user's container using a text editor.
For non-user mailboxes, the Sieve script exists in the maildir
as before, and although in theory can continue to be edited using
ManageSieve, I don't think this scenario has ever been possible,
and isn't now, since ManageSieve currently only facilitates editing
of personal Sieve rules for the authenticated user.

One caveat is that the symlink to the active Sieve script (or the
actual script, if it exists and isn't a symlink) remains in the
maildir. This is because symlinks contain the path of the target
as a string, and don't actually point to the inode of the target.
Thus, within the container, the symlink will not work if created
in the host, and vice versa. Additionally, showing a path on the
host within the container is undesirable, since this leaks information
to the user. As such, the symlink remains in the maildir. This means
users must use the ManageSieve script in order to change the active
script. On the upside, users won't be confused by a "weird symlink"
in the .configs directory, and they can still edit the scripts themselves
directly in the .configs directory.

Consequently, the MailScript changes are backwards-compatible,
but the Sieve changes are not fully backwards-compatible.

A few bugs from the previous commit (fef8df2b6c7c8ada47c2f751801ca22421dbf380)
have also been fixed.

There are a few edge cases currently:

* Non-user mailboxes (e.g. public/shared mailboxes) do not have
  users associated with them, and by extension, container environments.
  Thus, EXEC cannot be used with these mailboxes, except in global rules.
  This limitation could be worked around in the future.
* Environment variables like $HOME and shell shorthand like ~ are
  not evaluated prior to the program being launched, so full paths
  always need to be provided (BBS variables are okay). This isn't
  really specific to this change, but is important since it is
  intuitive to use such syntax in mailbox rules; currently, this doesn't work.