mod_smtp_delivery_local: Fix notifications being generated for wrong mailbox.
When processing incoming messages, the root maildir for the entire mailbox,
as opposed to the actual maildir of the message, was being passed to
mailbox_notify_new_message. In the common case of messages being delivered
to the INBOX, this would not make a difference, but if a filter moved it
to a different folder, the mailbox's root maildir was still used, resulting
in a "new message notification" for the INBOX, even if the new message was
in a different folder. This manifested as receiving an untagged EXISTS in
the INBOX even when there were no new messages. This is fixed to use the
correct, specific maildir of the message, so that the notification is
generated for the correct mailbox (such as using an untagged STATUS if
needed).

A test has also been added to capture this case.