mod_smtp_greylisting: Add ability to greylist messages in SMTP.
This adds the ability to greylist incoming email, a common technique
used to defer potentially spammy messages which results in "less
spam by deferring", with the idea that many spammers do not retry
temporarily failures.

The implementation here, while following RFC 6647 to some extent,
takes a different approach from conventional approaches to greylisting,
with the primary aim of reducing the chances that legitimate (ham,
or non-spam) mail is delayed for any reason. In particular, greylisting
is contingent about meeting two conditions:

* A minimum fail count. Many spammers deviate from the SMTP standards
  in way that commit an outright protocol violation or are, at the very
  least, suspicious. We already track this and tarpit senders that
  increase the fail count, but this is also a good sign that the message
  could be spam, and greylisting might make sense.
* A minimum spam score, as reported by SpamAssassin. This is less
  conventional, since typically greylisting is performed in order to
  deter potential spam to reduce system load, as spam filtering is a
  resource-intensive process. However, as our focus is less on
  high-throughput and more on effectiveness and convenience, it can
  make sense to analyze the message for spam, and greylist the message
  if it has a higher spam score.

Collectively, conditioning greylisting on these two criteria allows
us to avoid greylisting at all for the majority of legitimate mail.
This avoids delays that are commonly a source of frustration with
greylisting.